SDIT-3054: 👽️ Workaround for spring data jpa issue by returning java.lang.Integer instead

+++ src/main/java/uk/gov/justice/hmpps/prison/repository/jpa/repository/OffenceIndicatorRepository.kt
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository
 import uk.gov.justice.hmpps.prison.repository.jpa.model.OffenceIndicator
 
 interface OffenceIndicatorRepository : CrudRepository<OffenceIndicator, Long> {
-  fun deleteByIndicatorCodeAndOffenceCode(indicatorCode: String, offenceCode: String): Long
+  fun deleteByIndicatorCodeAndOffenceCode(indicatorCode: String, offenceCode: String): java.lang.Integer
   fun existsByIndicatorCodeAndOffenceCode(indicatorCode: String, offenceCode: String): Boolean
 }
